+++ entropylab/__init__.py
@@ -3,13 +3,13 @@ from entropylab.pipeline.api.data_reader import ExperimentReader
 from entropylab.pipeline.api.data_writer import RawResultData
 from entropylab.pipeline.api.execution import EntropyContext
 from entropylab.pipeline.api.graph import GraphHelper
-from entropylab.pipeline.api.param_store import ParamStore
 from entropylab.pipeline.graph_experiment import (
     Graph,
     PyNode,
     SubGraphNode,
     pynode,
 )
+from entropylab.pipeline.params.param_store import ParamStore
 from entropylab.pipeline.results_backend.sqlalchemy.db import SqlAlchemyDB
 from entropylab.pipeline.script_experiment import Script, script_experiment
 from entropylab.quam.core import QuAMManager, QuAM
